portfolio-analytics-agent/pyproject.toml
[project]
name = "portfolio-analytics-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.9.0",
    "google-adk>=1.4.2",
    "yfinance>=0.2.63",
]
